Reducer renaming and added project reducer
