+++ src/actions/index.js
@@ -19,6 +19,13 @@ export const addProject = (project) => {
 	}
 }
 
+export const pickProject = (project) => {
+	return {
+		type: 'PICK_PROJECT',
+		project
+	}
+}
+
 export const deleteProject = (id) => {
 	return {
 		type: 'DELETE_PROJECT',
