+++ src/reducers/index.js
@@ -1,9 +1,11 @@
 import { combineReducers } from 'redux';
 import palette from './palette-reducer';
+import projects from './projects-reducer';
 import project from './project-reducer';
 
 export const rootReducer = combineReducers({
 	palette,
+	projects,
 	project
 });
